Move tests to proper test suite
