+++ server/src/test/java/io/spine/server/aggregate/AggregatePartTest.java
@@ -24,35 +24,27 @@ import com.google.common.testing.NullPointerTester;
 import com.google.protobuf.Message;
 import io.spine.core.CommandEnvelope;
 import io.spine.server.BoundedContext;
-import io.spine.server.aggregate.given.AggregatePartTestEnv.AnAggregatePart;
 import io.spine.server.aggregate.given.AggregatePartTestEnv.AnAggregateRoot;
 import io.spine.server.aggregate.given.AggregatePartTestEnv.TaskDescriptionPart;
 import io.spine.server.aggregate.given.AggregatePartTestEnv.TaskDescriptionRepository;
 import io.spine.server.aggregate.given.AggregatePartTestEnv.TaskPart;
 import io.spine.server.aggregate.given.AggregatePartTestEnv.TaskRepository;
-import io.spine.server.entity.InvalidEntityStateException;
 import io.spine.test.aggregate.ProjectId;
 import io.spine.test.aggregate.Task;
 import io.spine.test.aggregate.command.AggAddTask;
-import io.spine.test.aggregate.user.User;
 import io.spine.testdata.Sample;
 import io.spine.testing.client.TestActorRequestFactory;
 import io.spine.testing.server.aggregate.AggregateMessageDispatcher;
 import io.spine.testing.server.model.ModelTests;
-import io.spine.validate.ConstraintViolation;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.lang.reflect.Constructor;
-import java.util.List;
 
 import static io.spine.base.Identifier.newUuid;
 import static io.spine.testing.DisplayNames.NOT_ACCEPT_NULLS;
-import static io.spine.testing.Verify.assertSize;
-import static io.spine.testing.server.entity.given.Given.aggregatePartOfClass;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.fail;
 
 /**
  * @author Illia Shepilov
@@ -106,49 +98,6 @@ class AggregatePartTest {
         taskRepository.store(taskPart);
         final Task task = taskDescriptionPart.getPartState(Task.class);
         assertEquals(TASK_DESCRIPTION, task.getDescription());
-    }
-
-    @SuppressWarnings("CheckReturnValue") // Method called to throw exception.
-    @Test
-    @DisplayName("throw InvalidEntityStateException if entity state is invalid")
-    void throwOnInvalidState() {
-        final User user = User.newBuilder()
-                              .setFirstName("|")
-                              .setLastName("|")
-                              .build();
-        try {
-            aggregatePartOfClass(AnAggregatePart.class)
-                    .withRoot(root)
-                    .withId(getClass().getName())
-                    .withVersion(1)
-                    .withState(user)
-                    .build();
-            fail("Should have thrown InvalidEntityStateException.");
-        } catch (InvalidEntityStateException e) {
-            List<ConstraintViolation> violations =
-                    e.getError()
-                     .getValidationError()
-                     .getConstraintViolationList();
-            assertSize(user.getAllFields()
-                           .size(), violations);
-        }
-    }
-
-    @Test
-    @DisplayName("update valid entity state")
-    void updateEntityState() {
-        User user = User.newBuilder()
-                        .setFirstName("Firstname")
-                        .setLastName("Lastname")
-                        .build();
-        AnAggregatePart part = aggregatePartOfClass(AnAggregatePart.class)
-                .withRoot(root)
-                .withId(getClass().getName())
-                .withVersion(1)
-                .withState(user)
-                .build();
-
-        assertEquals(user, part.getState());
     }
 
     private NullPointerTester createNullPointerTester() throws NoSuchMethodException {
