+++ testutil-server/src/test/proto/spine/testing/server/user.proto
@@ -19,12 +19,12 @@
 //
 syntax = "proto3";
 
-package spine.test.aggregate.user;
+package spine.testing.server;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.test.aggregate.user";
+option java_package = "io.spine.testing.server";
 option java_multiple_files = true;
 
 message User {
